+++ packages/editor/src/panels/VisualScript/controls/index.tsx
@@ -27,8 +27,8 @@ import React, { useState } from 'react'
 import { PiDownloadSimpleBold, PiPauseBold, PiPlayBold, PiTrashSimpleBold, PiUploadSimpleBold } from 'react-icons/pi'
 import { ControlButton, Controls } from 'reactflow'
 
-import { NodeSpecGenerator } from '@ir-engine/editor/src/components/visualScript/VisualScriptUIModule'
 import { GraphJSON, VariableJSON } from '@ir-engine/visual-script'
+import { NodeSpecGenerator } from '../../../components/VisualScript/VisualScriptUIModule'
 import { ClearModal } from '../modals/clear'
 import { HelpModal } from '../modals/help'
 import { Examples, LoadModal } from '../modals/load'
